# server/dist/swagger/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "Partial_User_": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "bvn_hash": {
                        "type": "string"
                    },
                    "verified": {
                        "type": "boolean"
                    },
                    "role": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "description": "Make all properties in T optional"
            },
...
            "PropertyDto": {
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Unique identifier for the property",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "title": {
                        "type": "string",
                        "description": "Property title",
                        "example": "3 Bedroom Luxury Apartment in Lekki"
                    },
                    "location": {
                        "type": "string",
                        "description": "Property location",
                        "example": "Lekki Phase 1, Lagos"
                    },
                    "price": {
                        "type": "number",
                        "format": "double",
                        "description": "Property price in Naira",
                        "example": 15000000
                    },
                    "type": {
                        "type": "string",
                        "description": "Property type",
                        "example": "apartment"
                    },
                    "images": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "description": "Array of property image URLs",
                        "example": [
                            "https://example.com/image1.jpg"
                        ]
                    },
                    "documents": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "description": "Array of property document URLs",
                        "example": [
                            "https://example.com/title-deed.pdf"
                        ]
                    },
                    "ai_valuation": {
                        "type": "number",
                        "format": "double",
                        "description": "AI-calculated property valuation",
                        "example": 16500000
                    },
                    "status": {
                        "type": "string",
                        "description": "Property status",
                        "example": "available"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Property creation timestamp",
                        "example": "2024-01-15T10:30:00.000Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Property last update timestamp",
                        "example": "2024-01-15T10:30:00.000Z"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "location",
                    "price",
                    "type",
                    "status"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "ErrorResponseDto": {
                "properties": {
                    "message": {
                        "type": "string",
                        "description": "Error message",
                        "example": "Invalid request parameters"
                    },
                    "statusCode": {
                        "type": "number",
                        "format": "double",
                        "description": "HTTP status code",
                        "example": 400
                    },
                    "name": {
                        "type": "string",
                        "description": "Error name/type",
                        "example": "ValidationError"
                    }
                },
                "required": [
                    "message",
                    "statusCode",
                    "name"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT token obtained from /auth/login"
            }
        }
    },
    "info": {
        "title": "coown-api",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {
            "name": "Salawu Olanrewaju Joseph"
        }
    },
    "paths": {
...
        "/users": {
            "get": {
                "operationId": "ListAll",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Partial_User_"
                                    },
                                    "type": "array"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": [
                                            {
                                                "id": "550e8400-e29b-41d4-a716-446655440000",
                                                "firstName": "John",
                                                "lastName": "Doe",
                                                "email": "john.doe@example.com",
                                                "role": "user",
                                                "verified": true,
                                                "isActive": true,
                                                "created_at": "2024-01-15T10:30:00.000Z",
                                                "updatedAt": "2024-01-15T10:30:00.000Z"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "LIST ALL USERS",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "role",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "isActive",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ]
            }
        },
        "/properties": {
            "get": {
                "operationId": "GetProperties",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PropertyDto"
                                    },
                                    "type": "array"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": [
                                            {
                                                "id": "550e8400-e29b-41d4-a716-446655440000",
                                                "title": "3 Bedroom Luxury Apartment in Lekki",
                                                "location": "Lekki Phase 1, Lagos",
                                                "price": 15000000,
                                                "type": "apartment",
                                                "images": [
                                                    "https://example.com/image1.jpg"
                                                ],
                                                "documents": [
                                                    "https://example.com/title-deed.pdf"
                                                ],
                                                "ai_valuation": 16500000,
                                                "status": "available",
                                                "created_at": "2024-01-15T10:30:00.000Z",
                                                "updatedAt": "2024-01-15T10:30:00.000Z"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Invalid filter parameters",
                                            "statusCode": 400,
                                            "name": "ValidationError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Properties"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "query",
                        "name": "location",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "apartment",
                                "house",
                                "land",
                                "commercial"
                            ]
                        }
                    },
                    {
                        "in": "query",
                        "name": "status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "available",
                                "under_contract",
                                "sold"
                            ]
                        }
                    }
                ]
            },
            "post": {
                "operationId": "CreateProperty",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropertyDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "title": "3 Bedroom Luxury Apartment in Lekki",
                                            "location": "Lekki Phase 1, Lagos",
                                            "price": 15000000,
                                            "type": "apartment",
                                            "images": [
                                                "https://example.com/image1.jpg"
                                            ],
                                            "documents": [
                                                "https://example.com/title-deed.pdf"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Property created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropertyDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Invalid property data",
                                            "statusCode": 400,
                                            "name": "ValidationError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Properties"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePropertyDtoType"
                            }
                        }
                    }
                }
            }
        },
...
        "/properties/{id}": {
            "get": {
                "operationId": "GetPropertyById",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PropertyDto"
                                        }
                                    ],
                                    "nullable": true
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "id": "550e8400-e29b-41d4-a716-446655440000",
                                            "title": "3 Bedroom Luxury Apartment in Lekki",
                                            "location": "Lekki Phase 1, Lagos",
                                            "price": 15000000,
                                            "type": "apartment",
                                            "images": [
                                                "https://example.com/image1.jpg"
                                            ],
                                            "documents": [
                                                "https://example.com/title-deed.pdf"
                                            ],
                                            "ai_valuation": 16500000,
                                            "status": "available",
                                            "created_at": "2024-01-15T10:30:00.000Z",
                                            "updatedAt": "2024-01-15T10:30:00.000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Property not found",
                                            "statusCode": 404,
                                            "name": "NotFoundError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Properties"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateProperty",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PropertyDto"
                                        }
                                    ],
                                    "nullable": true
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Property not found",
                                            "statusCode": 404,
                                            "name": "NotFoundError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Properties"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Partial_PropertyDto_"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteProperty",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Cannot delete property with active pools",
                                            "statusCode": 400,
                                            "name": "ValidationError"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Property not found",
                                            "statusCode": 404,
                                            "name": "NotFoundError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Properties"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/properties/{id}/valuation": {
            "get": {
                "operationId": "GetPropertyValuation",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "confidence": {
                                            "type": "string"
                                        },
                                        "estimatedValue": {
                                            "type": "number",
                                            "format": "double"
                                        }
                                    },
                                    "required": [
                                        "confidence",
                                        "estimatedValue"
                                    ],
                                    "type": "object"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "estimatedValue": 16500000,
                                            "confidence": "High"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseDto"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Property not found",
                                            "statusCode": 404,
                                            "name": "NotFoundError"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Properties"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}